refactor: remove Wallets component from HomePage
