+++ src/components/HomePage/index.ts
@@ -1,2 +1,2 @@
-import HomePage from './HomePage.container'
+import HomePage from './HomePage'
 export { HomePage }
