+++ src/components/Wallets/index.ts
@@ -1,0 +1,2 @@
+import Wallets from './Wallets.container'
+export { Wallets }
